Refactor Theme files (delete empty css files, replace inner directory, simplified structure) Add Theme header component
