+++ package.json
@@ -8,6 +8,7 @@
     "classnames": "^2.2.6",
     "lodash": "^4.17.10",
     "prop-types": "^15.6.2",
+    "query-string": "^6.1.0",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "react-router": "^4.3.1",
