--- src/Components/AnswerActions/AnswerActions.css
@@ -1,3 +1,0 @@
-.Answer__actions{
-
-}
