+++ src/Components/Program/SubjectList.jsx
@@ -5,7 +5,11 @@ import SubjectItem from './SubjectItem';
 class SubjectList extends Component {
     render () {
         return <div className="subjects-list">{
-            this.props.data.map((item, i) => <SubjectItem key={i} data={item} subjectLabel={item.subjectLabel} themesList={item.themesList} />)
+            this.props.data.map((item, i) => <SubjectItem
+                key={i}
+                data={item}
+                subjectLabel={item.subjectLabel}
+                themesList={item.themesList}/>)
         }
         </div>
     }
