+++ src/Components/Theme/CoinCount.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import './CoinCount.css';
 // import PropTypes from 'prop-types';
 
 export default class CoinCount extends React.Component {
