+++ src/Components/Theme/Theme.jsx
@@ -1,7 +1,8 @@
 import React, {Component} from 'react';
 import './Theme.css';
-import AnswerList from "../AnswerList/AnswerList";
-import QuestionList from "../QuestionList/QuestionList";
+import AnswerList from "./AnswerList";
+import ThemeHeader from "./ThemeHeader";
+import QuestionList from "./QuestionList";
 import Pagination from "../Pagination/pagination";
 import {paginate} from '../../util/paginate';
 import {getAnswers, getUsername} from '../../services/fakeAnswerService';
@@ -61,14 +62,7 @@ class Theme extends Component {
         return (
             <div className="Theme">
                 <div className="Theme-content">
-                    <div className="Theme-content__header">
-                        <h1 className="Theme-theme">Theme name</h1>
-                        <ul className="Theme-theme__types">
-                            <li className="Theme-theme__types__items"><a href="">Subject name</a></li>
-                            <li className="Theme-theme__types__items"><a href="">Level name</a></li>
-                        </ul>
-                    </div>
-
+                    <ThemeHeader/>
                     <AnswerList
                         answers={answersPage}
                         handleClick={this.handleQuestionClick}
